+++ primitives/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 use codec::{Decode, Encode};
+use frame_support::{parameter_types, BoundedVec};
 use scale_info::{prelude::vec::Vec, TypeInfo};
 #[cfg(feature = "std")]
 use serde::{Deserialize, Serialize};
@@ -9,13 +10,17 @@ use sp_runtime::{AccountId32, Perquintill, RuntimeDebug};
 
 pub mod traits;
 
+parameter_types! {
+	pub MaxHostLen: u8 = 255;
+	pub MaxDomainLen: u8 = 255;
+}
 pub const MILLICENTS: u128 = 100_000;
 pub const CENTS: u128 = 1_000 * MILLICENTS; // assume this is worth about a cent.
 pub const DOLLARS: u128 = 100 * CENTS;
 pub type ClusterId = H160;
 pub type DdcEra = u32;
 pub type BucketId = u64;
-pub type StorageNodePubKey = AccountId32;
+pub type StorageNodePubKey = AccountId32; // todo! to move to runtime
 
 // ClusterParams includes Governance non-sensetive parameters only
 #[derive(Clone, Encode, Decode, RuntimeDebug, TypeInfo, PartialEq)]
@@ -135,6 +140,67 @@ impl NodeModeFlags {
 	pub fn has_mode(&self, mode: NodeMode) -> bool {
 		(self.bits & (mode as u16)) != 0
 	}
+}
+
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Clone, Encode, Decode, RuntimeDebug, TypeInfo, PartialEq)]
+#[scale_info(skip_type_params(T))]
+pub struct StorageNode<T: frame_system::Config> {
+	pub pub_key: StorageNodePubKey,
+	pub provider_id: T::AccountId,
+	pub cluster_id: Option<ClusterId>,
+	pub props: StorageNodeProps,
+}
+
+impl<T: frame_system::Config> StorageNode<T> {
+	pub fn new(
+		node_pub_key: NodePubKey,
+		provider_id: T::AccountId,
+		node_params: NodeParams,
+	) -> Result<Self, NodeError> {
+		match node_pub_key {
+			NodePubKey::StoragePubKey(pub_key) => match node_params {
+				NodeParams::StorageParams(node_params) => Ok(StorageNode::<T> {
+					provider_id,
+					pub_key,
+					cluster_id: None,
+					props: StorageNodeProps {
+						mode: node_params.mode,
+						host: match node_params.host.try_into() {
+							Ok(vec) => vec,
+							Err(_) => return Err(NodeError::StorageHostLenExceedsLimit),
+						},
+						domain: match node_params.domain.try_into() {
+							Ok(vec) => vec,
+							Err(_) => return Err(NodeError::StorageDomainLenExceedsLimit),
+						},
+						ssl: node_params.ssl,
+						http_port: node_params.http_port,
+						grpc_port: node_params.grpc_port,
+						p2p_port: node_params.p2p_port,
+					},
+				}),
+			},
+		}
+	}
+}
+
+#[derive(Debug, PartialEq)]
+pub enum NodeError {
+	StorageHostLenExceedsLimit,
+	StorageDomainLenExceedsLimit,
+}
+
+#[cfg_attr(feature = "std", derive(Serialize, Deserialize))]
+#[derive(Clone, Encode, Decode, RuntimeDebug, TypeInfo, PartialEq)]
+pub struct StorageNodeProps {
+	pub host: BoundedVec<u8, MaxHostLen>,
+	pub domain: BoundedVec<u8, MaxDomainLen>,
+	pub ssl: bool,
+	pub http_port: u16,
+	pub grpc_port: u16,
+	pub p2p_port: u16,
+	pub mode: NodeModeFlags,
 }
 
 #[derive(Clone, Encode, Decode, RuntimeDebug, TypeInfo, PartialEq)]
